backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jalali_core==1.0.0
jdatetime==5.2.0
mysqlclient==2.2.7
pillow==11.3.0
PyJWT==2.10.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
gunicorn==21.2.0
